backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==4.2.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
